backend/requirements.txt
django==5.0.1
djangorestframework
psycopg2-binary
pillow